requirements.txt
Django>=4.2,<5.2
djangorestframework>=3.14,<3.16
django-cors-headers>=4.3,<4.6
requests>=2.31,<3.0
python-dotenv>=1.0,<2.0
gunicorn>=21.2,<23.0
whitenoise>=6.6,<7.0
